Make number of dummy elements grow as needed
Instead of being hard-coded to 16, the number of dummy elements will now be calculated based on the width of the page, and grow as necessary when the page is resized.

Note that elements are not destroyed, since even a very large monitor won't have too many elements and creating/destroying elements on every resize may introduce a (very minor) performance concern. This way its still fast, since at most it'll only ever create ~6 elements on most monitors, but it can still handle infinitely large monitors.